hasTrait -> hasTraits.
